Complete geoPHP method
This method ended up being a lot more efficient than the pointLocation
method.  It does require you to have the GEOS exentsion installed in
order to work. It also requires the use of another table (states) to
contain the shape data for the states. The ones that I was able to add
are stored in json/done.

GEOS: http://trac.osgeo.org/geos
I was unable to successfully install this extension on Windows, but I
was able to install it on Ubuntu 14.04.